core/statistics: statistics collection without a 'Collector'
This update removes the need to create a Collector for each action to
gather statistics. Instead, it allows direct calls to Statistics
methods, passing the action as an argument.

This change enables statistics collection for events across different
packages and functions without requiring a pre-created, shared Collector
for the action to be used by all functions collecting statistics.