warehouses/snowflake: implement the Identity Resolution
This commit implements the methods to perform Identity Resolution in
Snowflake. It is now possible to create a workspace that uses Snowflake
as a warehouse, import users, and execute Identity Resolution.

There are still limitations in various parts of the driver based on the
features that need to be implemented, as discussed in issue #842.

The 'normalize' function has been modified in this branch just to allow
for the development and testing of Identity Resolution. It needs to be
reviewed separately; see the issue #1121 for more information.

The correct handling of 'nil' values in the 'serializeValueToCSV'
function is discussed in the issue #1122.

The handling of Array properties in the alter schema has been done in
this PR; it may already be sufficient as is. However, this is discussed
in the issue #1123.

This commit also modifies the specified columns as 'AUTOINCREMENT' by
indicating a start, an increment, and the order, in order to obtain ID
generation similar to that of PostgreSQL.

For #842.