warehouses/postgresql: simplify and remove redundancy in test
This commit simplifies and removes redundancy in the 'Test_Merge' test
in the PostgreSQL driver, as it calls the 'typeToPostgresType' function
instead of hardcoding the name of the PostgreSQL type corresponding to
the Meergo type within the test. This is the same function we already
use to create columns in alter schema operations.