assets: fix and review the 'Check' button in warehouse settings page
Previously, the button simply pinged the warehouse. Now, however, it
calls the method "/warehouse/can-change-settings", which performs more
in-depth checks, so the simple term "Ping" is not accurate, as the ping
might succeed but other checks may not, subsequently showing the user
misleading error messages.

This commit therefore makes the action of the button more generic, using
the term "Check". It also modifies the error message accordingly.

In the future, we can find more specific terms and also add
documentation to the page for further clarity, perhaps indicating that
the data warehouse to be connected must be exactly the same as the one
already connected.