warehouses/snowflake: complete meergo-snowflake mapping, add tests
This change completes the mapping of Meergo types to Snowflake types and
adds tests for the Merge and Query methods, similar to those already
implemented for PostgreSQL. It also updates the README documentation
within the datastore package.

A follow-up commit will focus on normalizing the code between the
PostgreSQL and Snowflake drivers.

Fixes #1121
Fixes #1122
Fixes #1123