core/state: enforce listener call order and prevent removal
Previously, listeners were called in a random order, preventing any
listener from making assumptions about the internal state of other
components it might depend on, as their listeners may not have been
invoked yet. Listeners should instead be called in the order in which
they were added.

This change revises the listener logic to ensure that, once added,
listeners cannot be removed and will be called in the same order as they
were registered.