warehouses/snowflake: fix handling of arrays in Identity Resolution
This commit fixes the way arrays are handled in Identity Resolution in
Snowflake. In particular, it addresses the issue where NULL values in
identity arrays are incorrectly transformed into empty arrays for users.
This necessitated a workaround in the 'scanner.go' file.

By leveraging the property of the Snowflake function 'ARRAY_AGG', which
returns an empty array in the case of null, this commit:

* fixes the behavior
* removes the workaround, which is no longer necessary
* adds additional tests for Identity Resolution