core/transformers/mappings: avoid in-place mods for events & 'eq/ne'
Previously, transformations were always applied in-place. Specifically,
for types such as Object, Array, and Map, input values were modified
directly to produce output values. However, this approach posed risks
for events, as the original event would be altered, leading to potential
data race conditions and errors in other parts of the application where
the event is accessed. Additionally, in-place modifications occurred
even when expressions included the "eq" and "ne" functions.

This change ensures that in-place modifications are avoided for events
and when evaluating expressions containing the "eq" and "ne" functions.