cmd, core: remove obsolete calls to tracing functions
This commit removes obsolete calls to tracing functions. Such calls are
considered obsolete as their purpose was not to trace something
specific, but to trace "a bit of everything" within Meergo for testing
when we implemented the tracing framework.

Since the tracing framework is being reviewed/cleaned up by PR #1137,
this commit aims to clean up the calls to the framework method.

The more appropriate calls, capable of providing useful information,
will be added later.