core/transformers/mapping: disable layout in intermediate conversions
As was previously done for purpose, conversions performed during
intermediate steps should not use layout formatting for DateTime, Date,
and Time values. Formatting with a layout should only occur in the final
stage, when the evaluation result is converted to the destination type.
This update sets the layout to nil for intermediate conversions.