core: fix action schema validation
When the schema of a destination database action contains a property
with the same name as the action table key, it was being incorrectly
validated as if it were the table key, even if the property was not at
the first level of the schema.

This change fixes that issue.