assets: show "Go to overview" only if the error is bound to an execution
On the page showing the list of actions of a connection, when an action
is executed, in the event of an error, a tooltip is shown with the error
message and a "Go to overview" link which takes the user to the list of
executions of the connection, where the error is shown in more detail.

Prior to this commit, the "Go to overview" link was shown either when an
unprocessable error occurred in the call to the server, or when the call
ended without an error but an error was associated with the execution
itself.

This commit, on the other hand, ensures that the link is only shown if
the error is associated with the execution, and avoids showing it in the
case of an unprocessable error, as in that case the execution is not
even created and will therefore not be present in the list of executions
on the connection's overview page.

Close #1015.