meergo: support building and running of Meergo in a Docker container
This commit makes it possible to build a Docker image of Meergo and then
run that image in a container.

The procedure for doing this is documented in the README.md file,
section "Docker".

It is worth noting that the Dockerfile has been placed in the root of
the repository both because this seems to be a convention (but we can
revisit this when we further explore how others do it), and because,
when building, the Dockerfile cannot reference files that are located at
a higher level. Therefore, placing the Dockerfile in a subdirectory
would have required inventing a way, perhaps not so simple, to build
Meergo.