assets, cmd: keep 'meergo-assets' directory on the repository
Currently, running 'go vet' on a clean repository fails because the
files required for the embed "meergo-assets" are not present in the
repository.

This commit ensures that this directory always remains in the
repository, and that the files within it, ignored by Git, are managed
individually by Meergo, without removing and rebuilding the entire
directory, as is currently done on the main branch.