core/connectors: handle database connector errors as Unavailable
This change ensures that errors caused by a database connector returning
a column name that is not a valid property name, as well as errors
during row scanning, are converted to the UnavailableError type. This
allows the API to return a 503 Service Unavailable instead of a 500
Internal Server Error.

Fixes #1147