javascript-sdk: add `.gitignore` that ignores `node_modules`
This is necessary because, when running 'commit/commit.go', it is
expected that the repository remains clean and no modified files are
present.

This is currently not the case, as the files inside 'node_modules' are
being modified, which causes the final clean check in the GitHub Action
to fail.