assets: do not call the `hasOwnProperty` method on `ref` objects
This commit fixes an error where the `hasOwnProperty()` method was
called on `ref` objects to check whether a property belonged directly to
them or rather to the corresponding prototypes.

Since React `ref` objects are special objects that do not have an
associated prototype, the `hasOwnProperty()` method is not present on
them, and calling it caused a client internal error that led to issues
in the UI.

The commit therefore removes all invocations to `hasOwnProperty()` made
on `ref` objects, since they are not needed because they are objects
that do not inherit from a prototype, and so in this context, checking
whether a property is associated directly with the object or its
prototype is redundant.