.github/workflows: run full tests instead of passing `--short` opt
This commit modifies the GitHub Action that tests so that it runs the
full tests, not just the tests with '--short'.

This causes the Action to fail on 'main', as the tests cannot pass, but
this is necessary to proceed with the development of the PR that uses
testcontainers (#1146).

Since the GitHub Action is currently not executed automatically but only
manually when needed, this does not pose a problem.