assets: move the filters section in file storage's import actions
This commit changes the position of the filter section in the file
storage's import actions.

The section is now only shown after the file settings have been entered
and confirmed, just above the special properties section, as it is
necessary to have the input schema of the file in order to show the
correct options in the filter's `Property` combobox.

Close #1139.