cmd, test: make test work on HTTP instead of HTTPS
This commit makes the tests work over HTTP instead of HTTPS.

Using HTTPS is an unnecessary complication (since the purpose of the
tests is not to verify the correct functioning of HTTPS) and also causes
issues with certificates when running tests on GitHub Actions.

Note that to make the tests work, it was necessary to ensure that
cookies were Secure only when the server is HTTPS. This is related to
the fact that we do not have any other login mechanism. In this regard,
see issue https://github.com/open2b/meergo/issues/1153.