warehouses/snowflake: rename tables and columns to uppercase
This update renames the table and column names to uppercase.
Additionally, it ensures that table and column names are always quoted
in queries to prevent conflicts with Snowflake's reserved keywords.

This commit does not address keeping the Snowflake driver code aligned
with the PostgreSQL code, nor does it ensure that the 'merge.go' file
stays synchronized with the connector. These aspects will be handled in
subsequent commits.

Closes #842
Closes #1134