all: refactor the app writer and `Upsert` method for app connectors
This change refactors the app writer and the Upsert method for app
connectors. Previously, the connector received a single record to
process, even though the Upsert method can handle multiple records.

With this change, the connector receives all the queued records via the
new Records type, which implements a collection of records with methods
that cover all possible use cases for a connector. This allows for
optimized record exporting in batches.

Furthermore, Meergo no longer waits for the completion of one Upsert
call before making the next, further optimizing the export process for
apps.

Additionally, the RecordsError type now allows specifying exactly which
records failed, along with their respective errors.

Points not covered in this commit:

  * If a connector updates one record at a time, it might be more
    efficient to first read the records to be updated from the app and
    then only update those that have actually changed. This was handled
    in the previous version, but it didn’t account for the possibility
    that some property may be writable but not readable.

  * Currently, Meergo concurrently calls Upsert multiple times without
    considering whether ongoing calls might be waiting due to the
    backoff policy. A future commit will ensure that only one HTTP
    client is used per connector, shared with the dispatcher, and that
    Meergo takes into account the state of ongoing requests before
    starting new ones.