meergo: using generic container instead of `mysql` testcontainer
A generic container is used in MySQL tests instead of:

   github.com/testcontainers/testcontainers-go/modules/mysql

because 'testcontainers-go/modules/mysql' seems to not allow to
increase the startup timeout above 1 minute, causing the startup to
fail in cases when the host is under load (for example when executing
many tests).

Based on the tests conducted, it seems that issue #1159 is resolved. If
that is not the case, we will reopen it in the future.

Resolves #1159.