json,core/transformers: fix `Unmarshal` and remove `json.Unquote`
This update fixes the Unmarshal function in the transformers package and
removes the usage of the json.Unquote function, as the
'Value.AppendUnquote' method can be used instead.

Previously, the Unmarshal function incorrectly checked v to determine if
it was valid JSON, rather than verifying the value obtained after
unquoting v. This commit corrects that mistake by ensuring the proper
validation is performed.