meergo, test: make the test host configurable and document env vars
This commit makes possible to configure the host on which Meergo runs,
initiated by the tests, and also documents in the README all the
environment variables used by the tests in '/test/', along with their
descriptions and a default value.

As indicated in the README, the default values are generally sufficient
to run the tests, so it should not be necessary to set them.