core/datastore: fix user ordering in test
The users returned by Meergo, in the Identity Resolution tests, are
sorted solely by email, as it is only possible to sort users by one
property. Therefore, in the case of users with the same email but with
different values for first_name, etc..., the tests may randomly fail
based on how Meergo returned them.

For this reason, this commit sorts users based on all their Text
properties, in ascending order.