assets: move the linked connections to the actions page
This commit moves the linked connections from the connection's settings
to the connection's actions page.

In addition, the commit extends the `Section` component to support a
two-column layout, where the left column contains documentation related
to the section. This type of layout is now used by both the linked
connections and other parts of the UI.

Close #1016.