assets: handle errors during file reconfirmation
This commit updates the import action on file storage to handle errors
more consistently. If a file that was previously confirmed encounters
errors when "Preview" or "Confirm" is clicked again, the transformation
and filter section will be disabled. The user must then reconfirm the
file without errors to re-enable the sections and saving the action.

Close #1058.