all: make the user schema an API parameter instead of being hard-coded
This commit ensures that the user schema is no longer hardcoded within
the APIs and data warehouses, but is instead a parameter of the method
that adds a new workspace.

So now the drivers no longer know the default schema, and therefore
provide the option to initialize a warehouse with a schema provided by
Meergo.

Currently, the "externally passed" schema is the same as the one that
was hardcoded in the software, meaning it only contains a property
"email". However, since it is now defined in the UI, it is sufficient to
add new properties there for the workspaces to be initialized with a
richer schema. This will be done in subsequent commits.

For #1075.