test: simplify and speed-up a bit tests initialization
Since the method that creates a workspace now requires an initial
schema, this commit modifies the tests inside "/test" so that, instead
of creating a workspace with a minimal schema and then performing an
alter schema to populate it, it directly creates a workspace with the
schema used by the tests.

From the perspective of the tests, there is still the option, as it is
now on the main branch, to initialize the workspace with a minimal
schema (see the variable "minimalUserSchema").

For #1075.