assets: avoid displaying Bad Request errors directly to the user
This commit ensures that messages related to Bad Request errors are not
shown directly to the user inside the "toast", as these could confuse
them.

Instead, now the Bad Request error is logged in the console of the
browser and a more generic "An internal error occurred" message is shown
to the user.

Close #1180.