test/meergotester: make Meergo used in tests to serve the UI assets
This involves both the case when Meergo is started "embedded" in the
testing process (used when debugging tests) as well as the case when
Meergo is started in a separated process (the case when running more
than one test at once).

This commit is a prerequisite for implementing automated UI tests.