meergo, assets, tests: automatically run UI tests in reproducible env
This commit ensures that UI tests executed via Playwright, instead of
requiring an external environment to be set up and then a subsequent
manual execution, are run automatically when executing Go tests within a
reproducible environment.

Consequently, the documentation and infrastructure for the manual
procedure to run UI tests have also been removed, as it is no longer
recommended since the external environment must be reproducible.
Therefore, these tests should be initiated in a controlled environment,
which still maintains the simplicity and flexibility of manually
executed tests.

This commit also modifies some details in the test specifications to
adapt them for execution in the testing environment. In this regard,
please refer to the changes made to the file
'assets/tests/actions.spec.ts'.

With this change, UI tests can also be executed on GitHub Actions. There
are still modifications to be made to the Action to prevent the tests
from aborting in case they take too long, but this is not part of this
PR.

Resolves #1151.