assets: check if required property is in matching in mapping validation
This commit ensures that when the mapping of an action is validated, if
a property is required but has not been transformed, a check is made to
see if the property in question is in the matching before giving an
error.

This ensures that the UI does not return an unnecessary error because it
believes that the required property has not been used, when in fact it
has been used correctly in matching.

Close #1195.