assets: show confirmation on first mode switch during action editing
This commit ensures that the confirmation dialogue is not only shown
when trying to switch the mode after modifying the transformation, but
also when the transformation is not modified but the mode is being
changed for the first time during the editing of the action.

This ensures that when the page of an already added action is opened,
even if the transformation has not been changed, trying to switch the
mode will open the confirmation dialogue.

Close #1194.