commit: do a refactor of the CLI
This commit makes several changes to the command-line usage of the
'commit' script.

In particular:

* Improves documentation during command-line usage and error handling
* Fix the output of the "rm go.sum" command, which prints a bad path
* Moves parsing and command-line options handling into a separate
  function, for clarity and code isolation

With the changes made, script runs like:

    go run ./commit hello

now return a clear error instead of silently accepting the parameter
without it having any effect, creating cases of misunderstanding and
confusion in usage.

Additionally, the output of the command:

    go run ./commit --help

is now more complete and comprehensive.