commit: increase default test timeout to make GitHub Action execute
The GitHub Action that runs the tests could no longer complete execution
because the default 'go test' timeout of 10 minutes was not long enough
to run the tests inside '/test'. This commit therefore increases that
timeout to a reasonable value that should not stop the tests from
running under any circumstances.