all: set default displayed properties when adding a workspace from UI
This commit ensures that when a workspace is added from the UI, it has
the displayed properties set to a default.

To make this change, it was necessary to modify the API method that adds
the workspace, so that this method takes the displayed properties, just
as it does for the method that performs the update.

Resolves #1186.