core: improve action schema validation error messages
The error message for action schema validation was difficult to
interpret as it focused on the specific requirement that had been
violated. This change improves the clarity by specifying the schema
where the error is located and the path of the invalid property, along
with a brief description of what is invalid. Detailed requirements for
action schema properties should be referenced in separate documentation.

Additionally, it slightly adjusts the error messages returned by the
datastore and the user schema update process.