assets: identify mapping properties as required based on export mode
On the action page, an icon representing an asterisk is shown at the
mapping inputs, indicating whether the property being mapped to is
required.

Currently the asterisk is shown regardless of the export mode if the
mapped property is `createRequired` or `updateRequired`.

This commit, on the other hand, ensures the icon to be shown when the
property is `createRequired` only if the export mode contains `Create`
and when the property is `updateRequired` only if the export mode
contains `Update`, thus unifying with the rest of the logic in the UI.