core: transpose and refactor file `Actions.csv`
This commit transposes the CSV file Actions.csv, since we will be
increasing the number of information and fields, while the number of
action "types" will presumably remain fixed (at least for a while).

This therefore allows for greater scalability and readability of the
file.

This commit also refactors the file, adding "sections" inside it to
further increase readability.