assets: review the usage of "required" in transformations
This commit brings changes in the use of "required" in the
transformation of the action, both aesthetically and functionally.

The wording "required" has been standardized aesthetically in all places
where it is used, making it red and bold to accentuate its visibility.

Additionally, the "required" wording is now shown based on a more
thorough logic: it is always shown if it is on a first-level property,
or conditionally on other properties when one of their parent properties
or siblings are transformed.

As far as validation is concerned, we now require the user to pass a
property in the transformation (either in mappings or in the function)
when the property is "required" and is first-level, or when it is not
first-level but one of the siblings properties have been passed.