warehouses: make the Repair also correct the operations
This commit ensures that the Repair operation corrects the operations table,
both for PostgreSQL and Snowflake. This is necessary in situations where, for
example, Meergo have been abruptly interrupted during an operation.

Additionally, this commit removes old commented code that handled
"automatically" repair of operations, but did not work.

This change is intended as a solution to issue #1046, which will be
subsequently closed and replaced with a broader issue concerning a
refactoring of the code that manages the operations.

Note that Snowflake code testing will be done later.