all: remove the `deleted` param. from the `Merge` method of connectors
Since:

* in Meergo we do not pass any value for the "deleted" parameter to the
  'Merge' method of the database connectors, as we do not manage the
  deletion of users from database tables
* some connectors (e.g., MySQL) do not implement deletion, as is
  documented in the interface instead

This commit removes the "deleted" parameter from the interface, and
consequently from the code and the documentation. When we evaluate
whether to implement user deletion in the future, we will consider
adding this parameter back.

Of course, the parameter remains for those functions that are copy-paste
from the data warehouse, where we do use delete.

Also updates the documentation of the export, specifying that the
deletion is not supported.