all: clean up terms for workspace types and drivers
The term "type" is now used to reference a workspace type, while the
expanded form "workspace type" is used for clarity. The term "name" is
used specifically to refer to the name of the type.

For type implementations, the term "driver" is used, with the expanded
form "workspace driver" for clarity.

This update also removes the unused 'core.WarehouseType'.

Additionally, it updates the following endpoint:

GET /warehouses → GET /warehouse/type