core: fix validation of constant mapping and function transformations
Currently, the validation of mappings and transformation functions, when
these are constant (i.e., when there is no input property) is
inconsistent.

This commit makes the validation consistent with what is documented in
the Actions.csv file, and also updates some code that seems to have been
outdated for a long time.

It also adds additional tests on the action validation function.