core: rename `Connection.Records` to `File` and update parameters
This change renames the 'Connection.Records' method to 'File', swaps the
positions of the 'path' and 'format' parameters, and updates the
documentation to refer to 'format' instead of 'fileConnector'.

Additionally, it reorders the parameters in the 'Connection.Sheets'
method to align with the order in 'File' and 'Records', and updates its
documentation accordingly.