core: validate event type and target during action creation
It is not clear why nothing has been validated until now (whether due to
a simple error or to avoid introducing a dependency on the current state
of the connections and the actions that can be created).

However, the current implementation is partial and problematic because
it never performs the necessary checks.

This commit modifies the validation of the actions (during creation) and
updates the documentation accordingly.