all: remove the concept of "privacy region", leaving just in connectors
We have established that the privacy region should be removed from the
workspace and included as a specific setting for the connectors, where
necessary (see #1239).

So, this commit removes the privacy region and adds it as a setting in
the Mixpanel connector, which is currently the only one that used it.

Resolves #1239.