core: refactor `Connector` to separate source and destination settings
Previously, it was impossible to determine if a connector could only be
used as a source or a destination, or if the settings were supported
solely for one of these roles.

To resolve these issues, this change refactors the Connector type by
introducing 'Source' and 'Destination' fields, which contain properties
relevant to the connector when used in each respective role.

These fields allow for indicating whether a connector cannot be used as
a source or destination by assigning a nil value to the respective
field.

Additionally, this change moves the 'Connector.HasSettings' field into
both the 'Source' and 'Destination' fields, differentiating the support
for settings based on the role. The 'Connection.HasSettings' field is
also removed.

Fixes #1178