assets: review the position of the property caret in the testing mode
This commit moves the caret icon used for the expansion of the testing
mode properties to the left side of the property name (and to the left
side of the selection checkbox in the case of function-based
transformation).

The icon has been slightly revised in style, along with the padding of
the various testing mode panels.