assets: filter the formats shown during the creation of file actions
This commit ensures that when creating an action on a file storage, the
available file formats are filtered by taking into account their support
for use as source or destination.

This way, if, for example, a file format is not supported as source and
an import action is being created on a file storage, that format will
not be shown and cannot be selected by the user.

Close #1272.