doc: update Developers documentation after the refactor of connectors
This commit updates the Developers documentation after the refactor of
the connectors (done in the commit 42278714d6b4ad3705a089d10d3aac7890df7412).

This makes the documentation consistent with the effective
implementation of the connectors in Meergo and adds information about
the fact that now the connectors can more specifically implement only a
subset of the functionalities supported by Meergo.

After this commit, the examples in the documentation work again.