core/connectors: adjust schema values before returning connector schemas
This change allows a connector to return a source schema with
CreateRequired and UpdateRequired set to true, and a destination schema
with ReadOptional set to true.

The Schema and SchemaAsRole methods of App will automatically adjust
these values before returning the requested schema.

Additionally, this change updates the Schema and SchemaAsRole methods to
panic if the target is Events and the role is Source.