assets: revert "assets: set `updateRequired` to false in output schema"
This reverts commit 9c0e2cd23482cd7c3228208cfc542b3a0da67766.

Revert reason: UpdateRequired of the out matching property must have the
same value as in the destination schema to ensure alignment validation
does not fail. If the property is not present in the destination schema
(which can only happen if the export mode is UpdateOnly), the property
must be false, based on the type in the source schema.