core: move matching property removal and adjust `App.Writer` parameters
This change moves the logic that removes the out matching property from
the action's output schema, from the `App.Writer` method in the
`connectors` package to the `Action.exportUsers` method in the `core`
package.

Additionally, the parameters of the `App.Writer` method were changed to
pass the action field directly, instead of the entire action.