all: refactor validation of action target and event type
This commit refactors the validation of action target and event type,
which are messy and obsolete in some parts.

Regarding the target, the changes introduced by this commit are based on
the fact that the target should be validated only when the action is
created, as it cannot vary during action updates.

Regarding the event type, this commit rewrites the validation as the
event type should not be verified on the app itself, as it can vary at
any moment; moreover, as for the target, the event type cannot vary once
an action has been created, therefore, it does not need to be validated
every time the action is updated.

A character limit of 100 has been introduced for the event type ID.

This commit refactors the validations across the code, ensuring greater
formality, coherence and correctness, as well as updating the
documentation and the tests accordingly.