core: fix foreign key check in method `CreateWorkspace`
In case an organization does not exist anymore when creating a
workspace, the foreign key violation error is named
"workspaces_organization_fkey" instead of
"workspaces_keys_organization_fkey".
Therefore, the organization not found error is never reported and an
internal server error is returned by the APIs instead.

This commit fixes it.