assets, core: change default state of action enabling and scheduler
This commit ensures that, when a "batch" action is created, by default
it is enabled and the scheduler has no period.

As a result:

(1) it is sufficient with only one user click (on "Run now") to test the
action, without having to enable it first;
(2) in its initial state, the action only performs an operation if the
user explicitly interacts with it (the change to the scheduler keeps
this behavior, which is already present on 'main').

This does not apply to actions that import or export data from events.

This commit also updates the tests on the UI.