core: enable configurable incremental imports for actions
Previously, batch imports were always incremental for apps, and
incremental for databases and files only if the action included a
"last change time" column.

With this update, it's now possible to choose whether imports for a
source app, database, or file action should be incremental, regardless
of the action's configuration.

Additionally, the `/actions/:id/exec` endpoint has been modified to
support incremental imports based on the action's settings. The 'reload'
parameter has been renamed to incremental, allowing it to override the
action’s default setting.