connectors/postgresql: refactor the `Columns` method
Such method, on main, is still as it was when it was copy-pasted from
the warehouse. For that reason, it is old, confused, and have bugs.

This commit rewrites it, clearing and fixing it.

The file in which is declared will be renamed in a separate commit
to preserve the commit diff.

Note that the database export still has problems, which are independent
of the code changed by this commit. They will be fixed in next commits.

Resolves #1269.