core/connectors: relax `RecordWriter.RecordStrings` element count check
This change relaxes the `RecordWriter.RecordStrings` method to accept
records with fewer elements than the number of columns. Missing elements
are treated as empty strings.

This allows connectors, such as Excel, to call the method without
needing to first add elements to the record to match the column count.

Fixes #1290