core: when exporting to the database, require an additional column
This commit changes the export to databases, requiring that at least
another column is transformed in addition to the table key.

This avoids an inconsistency which is currently present with the
PostgreSQL and Snowflake database connectors, that return an internal
error if only the table key is transformed.

The issue #1298 concerns implementing this change in the UI.

The issue #1297 concerns supporting the export of just the table key
property (requires changes in some of the database connectors).