assets: do not close action page if server returns `ServiceUnavailable`
This commit ensures that during the opening and building of the action
page, if the server returns a `ServiceUnavailable` error in one of the
calls made to retrieve the action schemas, the action page is still
opened (albeit with the schemas-dependent sections, such as
transformation, disabled), so that the user can see the state of the
action and possibly make changes to fix the problem.

Close #1294.